dmm: 64bit compiler support
It was assuming address was 32bit and used u32 for it.
But u32 was defined as "unsigned long". It will be 32bit on
32bit compiler, but 64bit on 64bit compiler.
And it uses u32 for address. It is not good for 64bit compiler.
This patch fixed these issue.

One consern to the future is that mmap() requests offset as off_t
which is 32bit, and we are using address here.
It might be issue if physical address located on 64bit.

+++ dmm/cmd/cmd-me.c
@@ -35,7 +35,7 @@ static STHELPMSG s_Help[] = {
 static bool cmd( int nArgc, char *pstrArgv[] )
 {
 
-    u32           addr;
+    uintptr_t     addr;
     enum ESTYPE   stype;
     u32           data;
     u32           data_check  = 0xFFFFFFFF;
